+++ modules/i18n.nix
@@ -1,7 +1,8 @@
 {
   homeModule = { pkgs, lib, ... }:
     lib.mkIf pkgs.stdenv.hostPlatform.isLinux {
-      i18n.inputMethod.enabled = "fcitx5";
+      i18n.inputMethod.enable = true;
+      i18n.inputMethod.type = "fcitx5";
       i18n.inputMethod.fcitx5.addons =
         builtins.attrValues { inherit (pkgs) fcitx5-unikey; };